+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed:
+- Remove autofocus on measurement value field
+- Unify segmented time span controls on dashboard and habit pages
+- Increase analyzed habit completion time span to 90 days
+- Show date in habit completion rate chart tooltip
+
+## [0.8.205] - 2022-12-09
 ### Added:
 - Chart for habit completion rate over time